+++ module-system/module-implementations/sov-sequencer-registry/src/event.rs
@@ -1,3 +1,5 @@
+use sov_modules_api::Spec;
+
 /// Sample Event
 #[derive(
     borsh::BorshDeserialize,
@@ -8,9 +10,26 @@
     PartialEq,
     Clone,
 )]
-pub enum Event {
-    /// Sample event variant 1
-    Event1,
-    /// Sample event variant 2
-    Event2,
+pub enum Event<S: Spec> {
+    /// A sequencer was registered.
+    Registered {
+        /// The address of the sequencer that was registered.
+        sequencer: S::Address,
+        /// The amount of the initial deposit.
+        amount: u64,
+    },
+
+    /// A sequencer exited.
+    Exited {
+        /// The address of the sequencer that was exited.
+        sequencer: S::Address,
+    },
+
+    /// A sequencer deposited funds to stake.
+    Deposited {
+        /// The address of the sequencer that was deposited to.
+        sequencer: S::Address,
+        /// The amount of the deposit.
+        amount: u64,
+    },
 }